requirements.txt
numpy
faiss-cpu
tqdm
PyPDF2
streamlit
requests